Add an example test that uses PhantomJS
Added `spec/example.spec.html` which is an HTML fragment that is used
to run the test in `spec/example.spec.js` against the code in
`assets/scripts/example.js`. This required changing the code in
`example.js` to include some DOM manipulation, to ensure that PhantomJS
is really working.

Because `example.js` now does DOM manipulation, we don't want to require
it in `assets/scripts/index.js`, so I commented out that require. It was
just an example, so that shouldn't change anything.

Close #40.